multiple group deletes
